[FSDP][optim_state_dict] Enable cpu_offload config for optimzer state_dict
We had the option but never used cpu_offload as optimizer state_dict offloads the tensors to CPU by default. And this is usually most users want as the tensors are required to be moved to CPU eventually. However, we may want to disable offloading to CPU in some cases, epsecially for the debugging purpose. This PR lets optimizer state_dict read the flag.

Differential Revision: [D48913340](https://our.internmc.facebook.com/intern/diff/D48913340/)

ghstack-source-id: 199490000
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108434